Factored out file saving into its own module.
File saving has no dependence on the recognition engine now.
